# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "xinzhi-daoxue-api",
    "description": "芯智导学阶段 2.2 API。统一 Agent 注册、快速路由、受控降级与多星辰工作流共用同一 Provider 调用链。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/debug/agents/{agent_id}/validate": {
      "post": {
        "tags": [
          "agent-debug"
        ],
        "summary": "Validate Agent",
        "operationId": "validate_agent_api_v1_debug_agents__agent_id__validate_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Agent Api V1 Debug Agents  Agent Id  Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/agents/{agent_id}/mock": {
      "post": {
        "tags": [
          "agent-debug"
        ],
        "summary": "Run Mock",
        "operationId": "run_mock_api_v1_debug_agents__agent_id__mock_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDebugRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Mock Api V1 Debug Agents  Agent Id  Mock Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/agents/{agent_id}/compare": {
      "post": {
        "tags": [
          "agent-debug"
        ],
        "summary": "Compare Mock Cloud",
        "operationId": "compare_mock_cloud_api_v1_debug_agents__agent_id__compare_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCompareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Compare Mock Cloud Api V1 Debug Agents  Agent Id  Compare Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/agents/{agent_id}/contract-tests": {
      "post": {
        "tags": [
          "agent-debug"
        ],
        "summary": "Run Contract Tests",
        "operationId": "run_contract_tests_api_v1_debug_agents__agent_id__contract_tests_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Contract Tests Api V1 Debug Agents  Agent Id  Contract Tests Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentCompareRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 4000,
            "title": "Question",
            "default": "开发态契约测试输入"
          },
          "course_id": {
            "type": "string",
            "maxLength": 16,
            "title": "Course Id",
            "default": "CT"
          },
          "intent": {
            "$ref": "#/components/schemas/Intent",
            "default": "explain_concept"
          },
          "canonical_input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Canonical Input"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          },
          "allow_mock": {
            "type": "boolean",
            "title": "Allow Mock",
            "default": false
          },
          "allow_cloud": {
            "type": "boolean",
            "title": "Allow Cloud",
            "default": false
          },
          "cloud_sample": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Sample"
          }
        },
        "type": "object",
        "title": "AgentCompareRequest"
      },
      "AgentDebugRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 4000,
            "title": "Question",
            "default": "开发态契约测试输入"
          },
          "course_id": {
            "type": "string",
            "maxLength": 16,
            "title": "Course Id",
            "default": "CT"
          },
          "intent": {
            "$ref": "#/components/schemas/Intent",
            "default": "explain_concept"
          },
          "canonical_input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Canonical Input"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          },
          "allow_mock": {
            "type": "boolean",
            "title": "Allow Mock",
            "default": false
          }
        },
        "type": "object",
        "title": "AgentDebugRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Intent": {
        "type": "string",
        "enum": [
          "unknown",
          "solve_problem",
          "explain_concept",
          "verify_answer",
          "check_user_solution",
          "general_qa",
          "summarize_knowledge",
          "learning_advice"
        ],
        "title": "Intent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}